+++ CHANGELOG.md
@@ -17,10 +17,11 @@ OR
 Changelog
 =========
 ## 1.0.3 - TBD
+- Supports `delay` in Key Press action to add a 0.5 second delay.
 - Avoid losing config if exception during write.
 - Wait for Stream Deck to be attached on startup.
 - Drag and drop support for rearranging buttons around in UI.
-- Supports `plus` in the Key Press action to output `+`.
+- Supports `plus` and `comma` in the Key Press action to output `+` and `,` respectively.
 - Fixed `core dumped` error when closing.
 - Improved error handling for invalid command or key press actions.
 - Window title updated to `Stream Deck UI`.